refactored pure funcs into casper package
