+++ beacon-chain/casper/BUILD.bazel
@@ -1,0 +1,36 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "go_default_library",
+    srcs = [
+        "incentives.go",
+        "sharding.go",
+        "validator.go",
+    ],
+    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/casper",
+    visibility = ["//beacon-chain:__subpackages__"],
+    deps = [
+        "//beacon-chain/params:go_default_library",
+        "//beacon-chain/types:go_default_library",
+        "//beacon-chain/utils:go_default_library",
+        "//proto/beacon/p2p/v1:go_default_library",
+        "@com_github_ethereum_go_ethereum//common:go_default_library",
+        "@com_github_sirupsen_logrus//:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = [
+        "incentives_test.go",
+        "sharding_test.go",
+        "validator_test.go",
+    ],
+    embed = [":go_default_library"],
+    deps = [
+        "//beacon-chain/params:go_default_library",
+        "//beacon-chain/types:go_default_library",
+        "//beacon-chain/utils:go_default_library",
+        "//proto/beacon/p2p/v1:go_default_library",
+    ],
+)
