+++ beacon-chain/sync/BUILD.bazel
@@ -9,6 +9,7 @@ go_library(
         "//beacon-chain/types:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//shared/p2p:go_default_library",
+        "@com_github_ethereum_go_ethereum//event:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
     ],
 )
